Start moving heart drop to .h
